==> 2_Case_study_compounds.ipynb <==
# Case study compounds

Process file of EU-ToxRisk case study compounds assembed by Anne.

%run setup.py

### Config

# Directory for input files...

files_dir = 'files'

# Case study compounds file (TSV format) from Anne Hersey...

filename = 'case_study_compounds_141216.txt'

# Directory for reading and writing data files...

data_dir = 'data'

tick, cross = '\u2713', '\u2717'

### Initialisation

if not 'logger' in locals(): logger = make_logger.run(__name__)

### Load compounds

df0 = pd.read_csv(os.path.join(files_dir, filename), delimiter='\t')

df0.shape

df0.head(1)

# There are some issues with the compound names (i.e. column 'Compound')...

df = df0[['Compound', 'inchikey']].drop_duplicates()

duplicated = set(df['inchikey'][df['inchikey'].duplicated()])

df.query("inchikey in @duplicated").sort_values(['inchikey', 'Compound'])

# Drop unneeded columns and simplify column names...

df1 = (
    df0
        .drop(['Label', 'Smiles', 'deleted CAS Registry Numbers', 'alternative name'], axis=1)
        .rename(columns={
          'Compound':         'name'
        , 'canonical_smiles': 'smiles'
        })
        .fillna('')
    )

# Fix the compound name issues by standardising the names...

fixes = [
      ('Valproic acid', 'Valproic Acid')
    , ('hydroquinone',  'Hydroquinone')
]

for old, new in fixes: df1['name'] = df1['name'].str.replace(old, new)

# Check the compound name issues have disappeared...

df = df1[['name', 'inchikey']].drop_duplicates()

set(df['inchikey'][df['inchikey'].duplicated()])

df1.head(1)

# Pivot on the compounds...

df1['Case Study'] = ["Case Study {}".format(x) for x in df1['Case Study']] # Case study column name

df1['tick'] = tick

cs_cmpds_df = (
    df1.pivot_table(index=['CAS', 'name', 'smiles', 'inchi', 'inchikey'], columns='Case Study', values='tick', aggfunc='first', fill_value=cross)
    .reset_index()
)

cs_cmpds_df.columns.name = None

cs_cmpds_df.shape

cs_cmpds_df.head(1)

# Check there are no duplicated identifiers...

for col in ['smiles', 'name', 'inchi', 'inchikey']: assert not cs_cmpds_df[col].duplicated().sum()

# Check for dot-disconnected SMILES...

cs_cmpds_df[cs_cmpds_df['smiles'].str.contains('\.')]

The only dot-disconnected SMILES is for LPOS (Lithium PerfluoroOctane Sulfonate), CAS 29457-72-5.

# Add RDKit molecules for inspection...

PandasTools.AddMoleculeColumnToFrame(cs_cmpds_df, smilesCol='smiles', molCol='mol')

cs_cmpds_df[['name', 'mol']].head()

cs_cmpds_df.to_pickle(os.path.join(data_dir, 'case_study_compounds.pkl'))

# cs_cmpds_df.drop('mol', axis=1).to_excel('case_study_compounds_141216.xlsx', index=False)

# cs_cmpds_df = pd.read_pickle(os.path.join(data_dir, 'case_study_compounds.pkl'))

# cs_cmpds_df.shape
