==> 3_Tox21_compounds_vs_case_study_compounds.ipynb <==
# Overlap of Tox21 compounds in PubChem with EU-ToxRisk case study compounds

* See [here](./case_study_compounds.ipynb) for further details of the EU-ToxRisk case study compounds.

* See [here](./Tox21_compounds.ipynb) for further details of the PubChem compounds.

### Configuration and Setup

%run setup.py

import sqlalchemy

# Import ChEMBL database connection details
# N.B. This module simply defines a dict containing the data required for a connection, e.g...
# 
# chembl = dict(
#       driver = 'oracle'
#     , host   = 'XXX.YYY.ZZZ'
#     , port   =  1531
#     , dbname = 'AAA'
#     , user   = 'BBB'
#     , passwd = 'CCC'
# )

from chembl import chembl

### Config

# ChEMBL schema (version) to use...

schema = 'chembl_22'

# Directory for reading and writing data files...

data_dir = 'data'

### Initialisation

if not 'logger' in locals(): logger = make_logger.run(__name__)

engine = sqlalchemy.create_engine("{driver}://{user}:{passwd}@{host}:{port}/{dbname}".format(**chembl))

### Reload compounds

# Reload Tox21 compounds...

sid_mols_df = pd.read_pickle(os.path.join(data_dir, './sid_mols.pkl'))

sid_mols_df.shape

sid_mols_df.head(1)

# Load case-study compounds (from Anne)...

cs_cmpds_df = pd.read_pickle(os.path.join(data_dir, './case_study_compounds.pkl'))

cs_cmpds_df.shape

cs_cmpds_df.head(1)

case_study_inchikeys = set(cs_cmpds_df['inchikey'])

len(case_study_inchikeys)

tox21_inchikeys = set(sid_mols_df['inchikey'])

len(tox21_inchikeys)

chembl_inchikeys = set(x['inchikey'] for x in engine.execute("""
select distinct
  d.standard_inchi_key as inchikey
from
       {schema}.assays              a
  join {schema}.activities          b on a.assay_id = b.assay_id
  join {schema}.molecule_hierarchy  c on b.molregno = c.molregno
  join {schema}.compound_structures d on c.parent_molregno = d.molregno
where
  a.description like 'PubChem BioAssay: Tox21. %%'
""".format(schema=schema)).fetchall())

len(chembl_inchikeys)

case_study_and_tox21 = case_study_inchikeys.intersection(tox21_inchikeys)

len(case_study_and_tox21)

case_study_and_chembl = case_study_inchikeys.intersection(chembl_inchikeys)

len(case_study_and_chembl)

One compound in the case study compound set is not in the ChEMBL set...

case_study_and_tox21.difference(case_study_and_chembl)

Five compounds in the case study compound set is not in the ChEMBL set...

case_study_and_chembl.difference(case_study_and_tox21)

Of these five, four are only recorded in ChEMBL as having been tested in cell viability assays...

sql = """
select distinct
    d.standard_inchi_key as inchikey
  , e.src_compound_id
  , a.description
from
       chembl_22.assays              a
  join chembl_22.activities          b on a.assay_id = b.assay_id
  join chembl_22.molecule_hierarchy  c on b.molregno = c.molregno
  join chembl_22.compound_structures d on c.parent_molregno = d.molregno
  join chembl_22.compound_records    e on b.record_id = e.record_id
where
  a.description like 'PubChem BioAssay: Tox21. %%'
and
  d.standard_inchi_key in ({inchikeys})
"""

inchikeys = ('NXFQHRVNIOXGAQ-YCRREMRBSA-N', 'UIFFUZWRFRDZJC-SBOOETFBSA-N', 'USMZPYXTVKAYST-UHFFFAOYSA-N', 'XMAYWYJOQHXEEK-ZEQKJWHPSA-N')

pd.read_sql_query(sql.format(inchikeys=','.join("'{}'".format(x) for x in inchikeys)), engine)

Thus, presumably, for some reason they have not been tested in primary activity assays and thus do not appear in the Summary assays we are interested in here.

By contrast, the other inchikey, **`JQXXHWHPUNPDRT-WLSIYKJHSA-N`**, is recorded in ChEMBL as having been tested in numerous assays...

pd.read_sql_query(sql.format(inchikeys="'JQXXHWHPUNPDRT-WLSIYKJHSA-N'"), engine).shape

This compound is actually Rifampicim, and is discussed in more detail below.

### Rifampicin

The version of Rifampicin in the case study compound set is taken from ChEMBL (**`CHEMBL374478`**) and has InChI Key **`JQXXHWHPUNPDRT-WLSIYKJHSA-N`**...

cs_cmpds_df.query("inchikey == 'JQXXHWHPUNPDRT-WLSIYKJHSA-N'")['name'].tolist()

HTML('''<iframe src="https://www.ebi.ac.uk/chembl/compound/inspect/{chemblid}" width=1000 height=750></iframe>'''.format(chemblid='CHEMBL374478'))

However, the version of Rifampicin in PubChem, **`CID5458213`**, has InChI Key **`FZYOVNIOYYPUPY-UXKYPCFPSA-N`**...

sid_mols_df.query("inchikey == 'FZYOVNIOYYPUPY-UXKYPCFPSA-N'")['CID'].unique().tolist()

HTML('''<iframe src="https://pubchem.ncbi.nlm.nih.gov/compound/{cid}" width=1000 height=750></iframe>'''.format(cid=5458213))

There is thus a mismatch between the stuctures (and hence InChI Keys) for Rifampicin in the case study compound set and PubChem.

Essentially, the InChI Key for Rifampicin is **`JQXXHWHPUNPDRT-WLSIYKJHSA-N`** in ChEMBL (and hence the case-study file from Anne) and **`FZYOVNIOYYPUPY-UXKYPCFPSA-N`** in PubChem...

# Query Case Study compounds table and PubChem substance/compound table with 'ChEMBL' InChI Key...

inchikey = 'JQXXHWHPUNPDRT-WLSIYKJHSA-N'

cs_cmpds_df.query("inchikey == @inchikey").shape[0], sid_mols_df.query("inchikey == @inchikey").shape[0]

# Query Case Study compounds table and PubChem substance/compound table with 'PubChem' InChI Key...

inchikey = 'FZYOVNIOYYPUPY-UXKYPCFPSA-N'

cs_cmpds_df.query("inchikey == @inchikey").shape[0], sid_mols_df.query("inchikey == @inchikey").shape[0]

To 'fix' this, we update the case study compound table to use the 'PubChem' InChI Key instead of the 'ChEMBL' one...

cs_cmpds_df.loc[cs_cmpds_df['inchikey'] == 'JQXXHWHPUNPDRT-WLSIYKJHSA-N', 'inchikey'] = 'FZYOVNIOYYPUPY-UXKYPCFPSA-N'

cs_cmpds_df.query("name == 'rifampicin'")[['name', 'inchikey']]

# Query Case Study compounds table and PubChem substance/compound table with 'ChEMBL' InChI Key...

inchikey = 'JQXXHWHPUNPDRT-WLSIYKJHSA-N'

cs_cmpds_df.query("inchikey == @inchikey").shape[0], sid_mols_df.query("inchikey == @inchikey").shape[0]

# Query Case Study compounds table and PubChem substance/compound table with 'PubChem' InChI Key...

inchikey = 'FZYOVNIOYYPUPY-UXKYPCFPSA-N'

cs_cmpds_df.query("inchikey == @inchikey").shape[0], sid_mols_df.query("inchikey == @inchikey").shape[0]

# Subset and re-order column names for convenience...

cs_cmpds_df = cs_cmpds_df[['inchikey', 'name'] + cs_cmpds_df.columns[cs_cmpds_df.columns.str.contains("^Case Study \d+$")].tolist()]

cs_cmpds_df.head(1)

# Save the modified version of case study compounds file...

cs_cmpds_df.to_pickle(os.path.join(data_dir, 'case_study_compounds_2.pkl'))
