==> 4_Tox21_get_data.ipynb <==
# Tox21 get data

Get Tox21 data from PubChem _via_ the [PUG API](https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html).

%run setup.py

### Config

# PUG API URL to retrieve all SIDs associated with an AID...

sids_url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/assay/aid/{aid}/sids/JSON"

# PUG API URL to retrieve data associated with an AID and a list of SIDs...

data_url = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/assay/aid/{aid}/JSON?sid={sids}"

# Number of entities to request data for in each call to PUG API...

chunk_size = 500

# Endpoint columns (ordered)...

endpoint_cols = ['Summary', 'Activity', 'Potency (uM)', 'Efficacy (%)', 'Viability Activity', 'Viability Potency (uM)', 'Viability Efficacy (%)']

# Directory for reading and writing data files...

data_dir = 'data'

### Initialisation

if not 'logger' in locals(): logger = make_logger.run(__name__)

### Reload Tox21 Summary Assays

See [here](0_Tox21_assays.ipynb) for details.

summary_assays_df = pd.read_pickle(os.path.join(data_dir, 'tox21_summary_assays.pkl'))

summary_assays_df.shape

## Get Tox21 data

Get assay data for Tox21 summary assays.

# Get data for an AID for a list of SIDs...

def do_sids(aid, sids):

    response = requests.get(data_url.format(aid=aid, sids=','.join(str(x) for x in sids)))

    assert response.status_code == 200

    response = response.json()['PC_AssayContainer'][0]

    ### assay_df = pd.DataFrame([(x['tid'], x['name'], x['description'][0], x['type'], x['unit']) for x in response['assay']['descr']['results']], columns=['tid', 'name', 'description', 'type', 'unit'])
    
    assay_df = pd.DataFrame([(x['tid'], x['name']) for x in response['assay']['descr']['results']], columns=['tid', 'name'])

    def f(record):

        df = pd.DataFrame([(y, list(z.values())[0], list(z.keys())[0]) for y, z in [(x['tid'], x['value']) for x in record['data']]], columns=['tid', 'value', 'dtype'])

        df['sid'] = record['sid']

        return df

    return pd.concat(f(x) for x in response['data']).merge(assay_df, on='tid').sort_values(['sid', 'tid']).drop('tid', axis=1)

# Get data for an AID...

def do_aid(aid):
    
    response = requests.get(sids_url.format(aid=aid))

    assert response.status_code == 200

    sids = response.json()['InformationList']['Information'][0]['SID']

    n_sids = len(sids)

    logger.info("\tGot {} SIDs ({} chunks to do)...".format(n_sids, ceil(10001 / chunk_size)))

    dfs = []

    for n, chunk in enumerate((sids[i:i+chunk_size] for i in range(0, n_sids, chunk_size)), 1):

        logger.info("\tStarting chunk {}...".format(n))

        df = do_sids(aid, chunk)

        logger.info("\t...finished chunk: got {} records.".format(df.shape[0]))

        dfs.append(df)

    df = pd.concat(dfs)

    df['aid'] = aid

    return df

n_aids = summary_assays_df['AID'].size
    
dfs = []

logger.info("Starting to process AIDs: {} to do...".format(n_aids))

for n, aid in enumerate(summary_assays_df['AID'], 1):
    
    logger.info("Starting AID {} ({}/{})...".format(aid, n, n_aids))

    df = do_aid(aid)
    
    dfs.append(df)
    
    logger.info("... done AID {}.".format(aid))
    
logger.info("Finished.")

tox21_data_df = pd.concat(dfs).reset_index(drop=True)

tox21_data_df.shape

tox21_data_df.head()

# Save a copy of the raw data, as it takes a while to generate...

# tox21_data_df.to_pickle(os.path.join(data_dir, 'tox21_data_raw.pkl'))

# tox21_data_df = pd.read_pickle(os.path.join(data_dir, 'tox21_data_raw.pkl'))

# tox21_data_df.shape

# Rename and reorder columns...

tox21_data_df = (
    tox21_data_df
        .rename(columns={'aid': 'AID', 'sid': 'SID', 'name': 'endpoint'})
        [['AID', 'SID', 'endpoint', 'value', 'dtype']]
        .reset_index(drop=True)
)

tox21_data_df.shape

tox21_data_df.head()

See [here](0_Tox21_assays.ipynb#assay_endpoints) for more details on the standardisation of assay endpoints.

# Standardise endpoint (a.k.a. 'name')...

tox21_data_df['endpoint_original'] = tox21_data_df['endpoint']

tox21_data_df['endpoint'] = tox21_data_df['endpoint_original'].apply(lambda x: re.sub('^\w+', '', x).strip() if not re.search("nm|Viability|Source|Supplier", x) else x)

tox21_data_df.head()

# Inspect changes...

df = tox21_data_df.query("endpoint != endpoint_original").apply(lambda df: "{} -> {}".format(df['endpoint_original'], df['endpoint']), axis=1).value_counts().to_frame('count').reset_index()

df.head(20)

# Filter rows on the standardised endpoint...

tox21_data_df = tox21_data_df[tox21_data_df['endpoint'].isin(endpoint_cols)]

tox21_data_df.drop('endpoint_original', axis=1, inplace=True) # The original endpoint column is now superfluous

tox21_data_df.shape

tox21_data_df.head()

# Check number of assays is as expected...

tox21_data_df['AID'].unique().size

# Save filtered table of Tox21 assay data...

tox21_data_df.to_pickle(os.path.join(data_dir, 'tox21_data.pkl'))

# tox21_data_df = pd.read_pickle(os.path.join(data_dir, 'tox21_data.pkl'))

# tox21_data_df.shape
