==> 5_Tox21_reformat_data.ipynb <==
# Tox21 reformat data

Reformat Tox21 data for export and visualization ( _e.g._ in [DataWarrior](http://www.openmolecules.org/datawarrior/) ).

%run setup.py

from math import log10

from functools import partial

### Config

# Endpoint columns (ordered)...

endpoint_cols = ['Summary', 'Activity', 'Potency (uM)', 'Efficacy (%)', 'Viability Activity', 'Viability Potency (uM)', 'Viability Efficacy (%)']

# Categories for the hit-call enpoints...

summary_categories =  ['active agonist', 'inconclusive agonist', 'inconclusive agonist (cytotoxic)', 'inconclusive agonist (fluorescent)', 'active antagonist', 'inconclusive antagonist', 'inconclusive antagonist (cytotoxic)', 'inconclusive', 'inactive']

activity_categories = ['active agonist', 'inconclusive agonist',                                                                           'active antagonist', 'inconclusive antagonist',                                        'inconclusive', 'inactive']

# Directory for reading and writing data files...

data_dir = 'data'

# Final output file name...

outfile_name = 'tox21_data_for_eu-toxrisk.csv'

tick, cross = '\u2713', '\u2717'

### Initialisation

if not 'logger' in locals(): logger = make_logger.run(__name__)

# Utility function to reorder dataframe columns...

def insert_column_after(df, col0, col1):
    cols = df.columns.tolist()
    cols.remove(col1)
    cols.insert(cols.index(col0)+1, col1)
    return df[cols]

### Reload Tox21 data

See [here](4_Tox21_get_data.ipynb) for details of the data-retrieval procedure.

tox21_data_df = pd.read_pickle(os.path.join(data_dir, 'tox21_data.pkl'))

tox21_data_df.shape

tox21_data_df.head(1)

### Reformat data

Pivot data table to create columns for the individual endpoints...

index_cols = ['AID', 'SID']

pivoted_df = (
    tox21_data_df.pivot_table(index=index_cols, columns=['endpoint'], values=['value'], aggfunc='first')
     .reset_index()
     .fillna('')
)

pivoted_df.shape

pivoted_df.head(1)

# Reformat column headers...

n_index_cols = len(index_cols)

pivoted_df.columns = pivoted_df.columns.droplevel(1).tolist()[:n_index_cols] + pivoted_df.columns.droplevel(0).tolist()[n_index_cols:]

# Reorder columns...

pivoted_df = pivoted_df[index_cols + endpoint_cols]

pivoted_df.head(1)

###  Column types

Modify column types where necessary.

pivoted_df.dtypes

#### Numeric columns

Convert numeric ( _i.e._ the Potency and Efficacy ) columns from object to float...

def to_float(x):
    
    try:
        
        return np.float64(x)
        
    except ValueError as exception:
        
        return np.nan
    
for col in pivoted_df.columns[pivoted_df.columns.str.contains('|'.join(re.escape(x) for x in ['Potency (uM)', 'Efficacy (%)']))]:
    
    pivoted_df[col] = pivoted_df[col].apply(to_float)

#### Hit calls

Set the various hit-call columns to be ordered categorical variables, so that their (non-alphabetical) ordering is preseved...

pivoted_df['Summary'] = pivoted_df['Summary'].astype('category', categories=summary_categories, ordered=True)

for col in pivoted_df.columns[pivoted_df.columns.str.contains('Activity')]:
    
    pivoted_df[col] = pivoted_df[col].astype('category', categories=activity_categories, ordered=True)

pivoted_df.dtypes

### Data modification

Some data modification to enable easier visualization in DataWarrior.

# Add AC50 columns for the dose-response (i.e. 'Potency') columns...

potency_to_ac50 = lambda x: -log10(x/10**6)

pattern = re.escape('Potency (uM)')

for potency_col in pivoted_df.columns[pivoted_df.columns.str.contains(pattern)]:
    
    ac50_col = re.sub(pattern, 'AC50', potency_col)
        
    pivoted_df[ac50_col] = pivoted_df[potency_col].apply(potency_to_ac50)
    
    pivoted_df = insert_column_after(pivoted_df, potency_col, ac50_col)
    
pivoted_df.shape

pivoted_df.head(1)

# There are some excessively high and low Efficacy values, which can make visualization difficult...

max_efficacy, min_efficacy = 200.0, -200.0

efficacy_cols = pivoted_df.columns[pivoted_df.columns.str.contains(re.escape('Efficacy (%)'))]

for col in efficacy_cols:
    
    print("{:<25s}: {:4d} HIGH, {:4d} LOW".format(col, (pivoted_df[col] > max_efficacy).sum(), (pivoted_df[col] < min_efficacy).sum()))

# Trim the excessive Efficacy values..

trimmer = lambda x: partial(max, min_efficacy)(partial(min, max_efficacy)(x))

for col in efficacy_cols:
    
    pivoted_df[col] = pivoted_df[col].apply(trimmer)
    
for col in efficacy_cols:
    
    print("{:<25s}: {:4d} HIGH, {:4d} LOW".format(col, (pivoted_df[col] > max_efficacy).sum(), (pivoted_df[col] < min_efficacy).sum()))

# Prepend category code (a number) to string value of categorical variables to preserve ordering one exported as string...

hit_call_cols = ['Summary', 'Activity', 'Viability Activity']

for col in hit_call_cols:
    
    pivoted_df[col] = ["{}) {}".format(x, y) if type(y) == str else '' for x, y in zip(pivoted_df[col].cat.codes, pivoted_df[col])]

pivoted_df.head(1)

### Compound structures

Add compound structures (as SMILES).

See [here](1_Tox21_compounds.ipynb) for further details on the Tox21 compounds.

# Load SID-to-structure mapping table (keeping only required colummns)...

sid_mols_df = pd.read_pickle(os.path.join(data_dir, 'sid_mols.pkl')).drop(['inchi', 'mol'], axis=1)

sid_mols_df.shape

sid_mols_df.head(1)

Note that a small number of SIDs (187) do not have structures, and hence do not appear in the structure table 'sid_mols_df'...

df = pivoted_df.merge(sid_mols_df, on='SID')

df[df['inchikey'].isnull()]['SID'].unique().size

The pivoted data is thus joined to the structure table using an inner join; records without an associated structure are lost, but as there is no structure it is unlikely to be missed.

pivoted_df = pivoted_df.merge(sid_mols_df, on='SID', how='inner')

pivoted_df.shape

pivoted_df.head(1)

### Case study info

Add case study information.

See [here](Tox21_PubChem_vs_case_study_compounds.ipynb) for further details on the modified version of the case study compounds table.

# Load the modified version of the case study compounds table...

cs_cmpds_df = pd.read_pickle(os.path.join(data_dir, 'case_study_compounds_2.pkl'))

cs_cmpds_df.shape

cs_cmpds_df.head(1)

# Add an 'Any Case Study' column for convenience...

cs_cmpds_df['Any Case Study'] = tick

cs_cmpds_df = insert_column_after(cs_cmpds_df, 'name', 'Any Case Study')

pivoted_df = pivoted_df.merge(cs_cmpds_df, on='inchikey', how='left')

pivoted_df.shape

# Make the cases study compound name column less ambiguous...

pivoted_df.rename(columns={'name': 'Compound Name'}, inplace=True)

# Fix missing values introduced by joins...

pivoted_df['Compound Name'].fillna('', inplace=True)

for col in pivoted_df.columns[pivoted_df.columns.str.match('^(?:Any Case Study|Case Study \d+)$')]:
    
    pivoted_df[col].fillna(cross, inplace=True)    

# Add a numeric version of 'Any Case Study' to allow sizing by case study membership in DataWarrior...

pivoted_df['ACS'] = pivoted_df['Any Case Study'].apply(lambda x: 1 if x == tick else 0)

pivoted_df.head(1)

### Assay information

See [here](0_Tox21_assays.ipynb) for further details on the Tox21 summary assays.

summary_assays_df = pd.read_pickle(os.path.join(data_dir, 'tox21_summary_assays.pkl'))

summary_assays_df.shape

summary_assays_df.head(1)

pivoted_df = pivoted_df.merge(summary_assays_df, on='AID')

pivoted_df.shape

# Prettify column names...

pivoted_df.rename(columns={'assay_name': 'Assay Name', 'target': 'Target', 'gene_id': 'Gene ID'}, inplace=True)

# reorder columns...

pivoted_df = insert_column_after(pivoted_df, 'AID', 'Assay Name')

pivoted_df.head(1)

# Sort first on assay name, as this is likely how the data will be browsed...

pivoted_df.sort_values(['Assay Name', 'SID'], inplace=True)

### Export data 

Export reformatted data as CSV file for import into DataWarrior.

pivoted_df.to_csv(os.path.join(data_dir, outfile_name), index=False)

## Examining hit-call counts

A brief digression to look at the distribution of the various hit calls.

Independent counts for various hit calls...

pivoted_df['Summary'].value_counts().to_frame('N').reset_index(drop=False).rename(columns={'index': 'Summary'}) # .sort_values('Summary')

pivoted_df['Activity'].value_counts().to_frame('N').reset_index(drop=False).rename(columns={'index': 'Activity'}) # .sort_values('Activity')

pivoted_df['Viability Activity'].value_counts().to_frame('N').reset_index(drop=False).rename(columns={'index': 'Viability Activity'}) # .sort_values('Viability Activity')

Contingent counts for various hit calls...

df = pivoted_df[hit_call_cols].copy()

df['count'] = '' # NB Add dummy column for group-by count

df = df.groupby(hit_call_cols).count().dropna() # NB Use of categorical variables introduces empty cells

df['count'] = df['count'].astype(int) # Introduction of empty cells coerces column to float

df

# Alternative view of contingent counts...

hit_call_cols_2 = list(reversed(hit_call_cols))

df = pivoted_df[hit_call_cols_2].copy()

df['count'] = '' # NB Add dummy column for group-by count

df = df.groupby(hit_call_cols_2).count().dropna() # NB Use of categorical variables introduces empty cells

df['count'] = df['count'].astype(int) # Introduction of empty cells coerces column to float

df
